+++ Coimbra.Editor/ProjectSettingsProviderRegister.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using UnityEditor;
-using UnityEngine;
 
 namespace Coimbra.Editor
 {
@@ -18,14 +17,10 @@ namespace Coimbra.Editor
             {
                 foreach (Type type in TypeCache.GetTypesWithAttribute<ProjectSettingsAttribute>())
                 {
-                    if (dictionary.ContainsKey(type))
+                    if (dictionary.ContainsKey(type) || !typeof(ScriptableSettings).IsAssignableFrom(type) || type.IsGenericType || type.IsAbstract)
                     {
                         continue;
                     }
-
-                    Debug.Assert(!type.IsAbstract, $"{nameof(ProjectSettingsAttribute)} should not be used on abstract type {type.FullName}!");
-                    Debug.Assert(!type.IsGenericType, $"{nameof(ProjectSettingsAttribute)} should not be used on generic type {type.FullName}!");
-                    Debug.Assert(typeof(ScriptableSettings).IsAssignableFrom(type), $"{type.FullName} should be a {nameof(ScriptableSettings)} to use {nameof(ProjectSettingsAttribute)}!");
 
                     ProjectSettingsAttribute attribute = type.GetCustomAttribute<ProjectSettingsAttribute>();
                     string path = $"{attribute.WindowPath}/{attribute.NameOverride ?? CoimbraEditorGUIUtility.ToDisplayName(type.Name)}";