+++ Coimbra/ProjectSettingsAttribute.cs
@@ -8,7 +8,7 @@ namespace Coimbra
     /// <summary>
     /// Apply this attribute to a <see cref="ScriptableSettings"/> that should show up in the project settings window.
     /// </summary>
-    [AttributeUsage(AttributeTargets.Class)]
+    [AttributeUsage(AttributeTargets.Class, Inherited = false)]
     [BaseTypeRequired(typeof(ScriptableSettings))]
     public sealed class ProjectSettingsAttribute : Attribute
     {