+++ Roslyn~/Coimbra.Roslyn/CoimbraTypes.cs
@@ -8,6 +8,10 @@
 
         public static readonly TypeString ManagedPoolClass = new("ManagedPool", Namespace);
 
+        public static readonly TypeString PreferencesAttribute = new("PreferencesAttribute", Namespace);
+
+        public static readonly TypeString ProjectSettingsAttribute = new("ProjectSettingsAttribute", Namespace);
+
         public static readonly TypeString ScriptableSettingsClass = new("ScriptableSettings", Namespace);
 
         public static readonly TypeString SharedManagedPoolAttribute = new("SharedManagedPoolAttribute", Namespace);